minor typecasting fixes

+++ src/main.rs
@@ -43,7 +43,7 @@ fn add_example_variables(server: &mut Server, ns: u16) {
 
         let _ = address_space.add_variables(
             vec![
-                Variable::new(&v1_node, "v1", "v1", 0 as i32),
+                Variable::new(&v1_node, "v1", "v1", 0i32),
                 Variable::new(&v2_node, "v2", "v2", false),
                 Variable::new(&v3_node, "v3", "v3", UAString::from("")),
                 Variable::new(&v4_node, "v4", "v4", 0f64),
@@ -107,7 +107,7 @@ fn add_example_variables(server: &mut Server, ns: u16) {
             data.1 = !data.1;
             let mut address_space = address_space.write();
             let now = DateTime::now();
-            let _ = address_space.set_variable_value(v1_node.clone(), data.0 as i32, &now, &now);
+            let _ = address_space.set_variable_value(v1_node.clone(), data.0, &now, &now);
             let _ = address_space.set_variable_value(v2_node.clone(), data.1, &now, &now);
         });
     }
